requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow-cpu==2.10.0
numpy==1.23.5
Pillow==10.1.0
python-jose==3.3.0
aiofiles==23.2.1
jinja2==3.1.2